refactor: extract idea seed list component
